@@ -16,20 +16,20 @@
 	"devDependencies": {
 		"@iconify/svelte": "^3.1.6",
 		"@sveltejs/adapter-static": "^3.0.1",
-		"@sveltejs/kit": "^2.0.0",
-		"@sveltejs/vite-plugin-svelte": "^3.0.0",
+		"@sveltejs/kit": "^2.5.27",
+		"@sveltejs/vite-plugin-svelte": "^4.0.0",
 		"@tailwindcss/vite": "^4.0.0",
 		"@tauri-apps/cli": ">=2.0.0-beta.0",
 		"prettier": "^3.2.5",
-		"prettier-plugin-svelte": "^3.2.2",
+		"prettier-plugin-svelte": "^3.2.6",
 		"prettier-plugin-tailwindcss": "^0.6.11",
-		"svelte": "^4.2.19",
-		"svelte-check": "^3.6.0",
+		"svelte": "^5.0.0",
+		"svelte-check": "^4.0.0",
 		"tailwindcss": "^4.0.0",
 		"tailwindcss-palette-generator": "^0.10.0",
 		"tslib": "^2.4.1",
-		"typescript": "^5.0.0",
-		"vite": "^5.1.5"
+		"typescript": "^5.5.0",
+		"vite": "^5.4.4"
 	},
 	"type": "module",
 	"dependencies": {
@@ -43,7 +43,7 @@
 		"@tauri-apps/plugin-process": "^2.0.0",
 		"@tauri-apps/plugin-shell": "^2.0.0",
 		"@tauri-apps/plugin-updater": "^2.0.0",
-		"bits-ui": "0.19.6",
+		"bits-ui": "^2.8.9",
 		"rehype-highlight": "^7.0.0",
 		"rehype-raw": "^7.0.0",
 		"svelte-exmarkdown": "^3.0.5"
