@@ -1,7 +1,12 @@
 <script lang="ts">
-	export let label;
-	export let description;
-	export let required = false;
+	type Props = {
+		label: any;
+		description: any;
+		required?: boolean;
+		children?: import('svelte').Snippet;
+	};
+
+	let { label, description, required = false, children }: Props = $props();
 </script>
 
 <div class="my-1 invalid:mx-5">
@@ -13,5 +18,5 @@
 	</div>
 	<p class="text-primary-300 pb-1 pl-0.5">{description}</p>
 
-	<slot />
+	{@render children?.()}
 </div>
