@@ -1,29 +1,34 @@
 <script lang="ts">
-	import { createEventDispatcher } from 'svelte';
+	import type { HTMLInputAttributes } from 'svelte/elements';
 
-	export let value: string = '';
-	export let size: 'sm' | 'md' | 'lg' = 'md';
+	type Props = {
+		value?: string;
+		size?: 'sm' | 'md' | 'lg';
+		class?: string;
+		onsubmit?: (value: string) => void;
+		onchange?: (value: string) => void;
+	} & Omit<HTMLInputAttributes, 'size' | 'onchange'>;
 
-	let className = '';
-
-	export { className as class };
-
-	const dispatch = createEventDispatcher<{
-		submit: string;
-		change: string;
-	}>();
+	let {
+		value = $bindable(''),
+		size = 'md',
+		class: className = '',
+		onsubmit,
+		onchange,
+		...props
+	}: Props = $props();
 </script>
 
 <input
 	type="text"
 	bind:value
 	autocomplete="off"
-	{...$$restProps}
-	on:keydown={(evt) => {
+	onkeydown={(evt) => {
 		if (evt.key === 'Enter') {
-			dispatch('submit', value);
+			onsubmit?.(value);
 		}
 	}}
-	on:change={() => dispatch('change', value)}
+	{...props}
+	onchange={() => onchange?.(value)}
 	class="{className} valid:focus:ring-accent-500 disabled:text-primary-400 bg-primary-900 text-primary-300 placeholder-primary-400 hover:ring-primary-500 min-w-0 grow rounded-lg px-3 py-1 invalid:ring-2 invalid:ring-red-500 hover:ring-1 focus:ring-2 focus:outline-hidden disabled:cursor-not-allowed text-{size} placeholder:text-{size}"
 />
