@@ -4,13 +4,17 @@
 
 	import ResetConfigButton from './ResetConfigButton.svelte';
 
-	export let entryId: ConfigEntryId;
-	export let locked: boolean;
+	type Props = {
+		entryId: ConfigEntryId;
+		locked: boolean;
+	};
+
+	let { entryId, locked }: Props = $props();
 
 	// we don't use hashtags in cfg files, since they are turned into comments
 	// but the color input works with the # prefix, so we have to separate them
 	let content = entryId.entry.value.content as string;
-	let hexCode = `#${content}`;
+	let hexCode = $state(`#${content}`);
 
 	async function onReset(value: ConfigValue) {
 		hexCode = `#${value.content}`;
@@ -27,5 +31,5 @@
 	}
 </script>
 
-<input type="color" class="grow" disabled={locked} bind:value={hexCode} on:change={submit} />
+<input type="color" class="grow" disabled={locked} bind:value={hexCode} onchange={submit} />
 <ResetConfigButton {entryId} {onReset} {locked} />
