@@ -3,11 +3,15 @@
 	import type { ConfigEntryId, ConfigNum, ConfigValue } from '$lib/models';
 	import ResetConfigButton from './ResetConfigButton.svelte';
 
-	export let entryId: ConfigEntryId;
-	export let locked: boolean;
+	type Props = {
+		entryId: ConfigEntryId;
+		locked: boolean;
+	};
+
+	let { entryId, locked }: Props = $props();
 
 	let value = entryId.entry.value;
-	let content = value.content as ConfigNum;
+	let content = $state(value.content as ConfigNum);
 	let type = value.type as 'int' | 'float';
 
 	function onReset(value: ConfigValue) {
@@ -24,7 +28,7 @@
 	disabled={locked}
 	step={type === 'int' ? 1 : 'any'}
 	bind:value={content.value}
-	on:change={submit}
+	onchange={submit}
 	class="focus:ring-accent-500 bg-primary-900 text-primary-300 placeholder-primary-400 hover:ring-primary-500 w-full grow rounded-lg px-3 py-1 hover:ring-1 focus:ring-2 focus:outline-hidden"
 />
 
