@@ -2,11 +2,15 @@
 	import Icon from '@iconify/svelte';
 	import { page } from '$app/stores';
 
-	export let to: string;
-	export let icon: string;
-	export let label: string;
+	type Props = {
+		to: string;
+		icon: string;
+		label: string;
+	};
 
-	$: active = $page.url.pathname === to;
+	let { to, icon }: Props = $props();
+
+	let active = $derived($page.url.pathname === to);
 </script>
 
 <a
@@ -16,5 +20,4 @@
 		: 'text-primary-500 hover:bg-primary-800 hover:text-primary-400'}"
 >
 	<Icon class="text-2xl" {icon} />
-	<!--<span class="hidden lg:block {active ? 'font-semibold' : 'font-medium'}">{label}</span>-->
 </a>
