@@ -1,12 +1,16 @@
 <script lang="ts">
 	import Icon from '@iconify/svelte';
 
-	let className = '';
+	type Props = {
+		class?: string;
+	};
 
-	export { className as class };
+	let { class: className = '' }: Props = $props();
 </script>
 
-<div class="{className} flex items-center rounded-lg bg-primary-700 py-1.5 pr-1 pl-3 text-primary-300">
+<div
+	class="{className} bg-primary-700 text-primary-300 flex items-center rounded-lg py-1.5 pr-1 pl-3"
+>
 	<Icon icon="mdi:lock" class="mr-2 text-xl" />
 	Profile is locked
 </div>
